+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2015-12-23 - 1.4.0
+------------------
+* Private methods are now protected.
+* Methods `Command::rules()` and `Command::permissions()` are now abstract.
+* Improve phpdoc.
+* Refactoring.
+
 2015-06-19 - 1.3.1
 ------------------
 * Added auto invalidate cache if use `yii\rbac\DbManager`.